Keep startup regressions under Swift file budget
Move the issue-3337 startup regression coverage out of WorkspaceUnitTests so the workflow guard continues enforcing the existing file-length budget after the merge from main.

Constraint: workflow-guard-tests enforces per-file Swift length budgets

Rejected: Raise the WorkspaceUnitTests budget | would weaken the guard for a test-only organization issue

Confidence: high

Scope-risk: narrow

Tested: python3 scripts/swift_file_length_budget.py --budget .github/swift-file-length-budget.tsv

Tested: git diff --check

Not-tested: Local XCTest suite; project policy runs tests in CI
