+++ cmuxTests/WorkspaceUnitTests.swift
@@ -667,56 +667,6 @@ final class KeyboardShortcutSettingsFileStoreTests: XCTestCase {
         XCTAssertEqual(store.activeSourcePath, settingsFileURL.path)
     }
 
-    func testSettingsFileStoreParsesNumberedShortcutWithoutConsultingActiveShortcutStore() throws {
-        let directoryURL = try makeTemporaryDirectory()
-        defer { try? FileManager.default.removeItem(at: directoryURL) }
-
-        let activeSettingsFileURL = directoryURL.appendingPathComponent("active.json", isDirectory: false)
-        try writeSettingsFile(
-            """
-            {
-              "shortcuts": {
-                "openBrowser": "cmd+3"
-              }
-            }
-            """,
-            to: activeSettingsFileURL
-        )
-
-        KeyboardShortcutSettings.settingsFileStore = KeyboardShortcutSettingsFileStore(
-            primaryPath: activeSettingsFileURL.path,
-            fallbackPath: nil,
-            startWatching: false
-        )
-        XCTAssertEqual(
-            KeyboardShortcutSettings.shortcut(for: .openBrowser),
-            StoredShortcut(key: "3", command: true, shift: false, option: false, control: false)
-        )
-
-        let parsedSettingsFileURL = directoryURL.appendingPathComponent("parsed.json", isDirectory: false)
-        try writeSettingsFile(
-            """
-            {
-              "shortcuts": {
-                "selectWorkspaceByNumber": "cmd+7"
-              }
-            }
-            """,
-            to: parsedSettingsFileURL
-        )
-
-        let store = KeyboardShortcutSettingsFileStore(
-            primaryPath: parsedSettingsFileURL.path,
-            fallbackPath: nil,
-            startWatching: false
-        )
-
-        XCTAssertEqual(
-            store.override(for: .selectWorkspaceByNumber),
-            StoredShortcut(key: "1", command: true, shift: false, option: false, control: false)
-        )
-    }
-
     func testSettingsFileStoreParsesRightSidebarShortcutBindings() throws {
         let directoryURL = try makeTemporaryDirectory()
         defer { try? FileManager.default.removeItem(at: directoryURL) }
@@ -922,112 +872,6 @@ final class KeyboardShortcutSettingsFileStoreTests: XCTestCase {
         )
 
         XCTAssertEqual(defaults.string(forKey: AppIconSettings.modeKey), AppIconMode.automatic.rawValue)
-        XCTAssertEqual(startObservationCallCount, 0)
-        XCTAssertEqual(stopObservationCallCount, 0)
-        XCTAssertEqual(imageRequestCount, 0)
-        XCTAssertEqual(runtimeIconSetCount, 0)
-        XCTAssertEqual(dockTileNotificationCount, 0)
-    }
-
-    func testSettingsFileStoreRestoresAbsentAppIconBackupDuringStartupWithoutTouchingAppKit() throws {
-        let defaults = UserDefaults.standard
-        let previousMode = defaults.object(forKey: AppIconSettings.modeKey)
-        let previousAppearance = defaults.object(forKey: AppearanceSettings.appearanceModeKey)
-        let previousBackups = defaults.data(forKey: settingsFileBackupsDefaultsKey)
-        defer {
-            if let previousMode {
-                defaults.set(previousMode, forKey: AppIconSettings.modeKey)
-            } else {
-                defaults.removeObject(forKey: AppIconSettings.modeKey)
-            }
-
-            if let previousAppearance {
-                defaults.set(previousAppearance, forKey: AppearanceSettings.appearanceModeKey)
-            } else {
-                defaults.removeObject(forKey: AppearanceSettings.appearanceModeKey)
-            }
-
-            if let previousBackups {
-                defaults.set(previousBackups, forKey: settingsFileBackupsDefaultsKey)
-            } else {
-                defaults.removeObject(forKey: settingsFileBackupsDefaultsKey)
-            }
-        }
-
-        defaults.removeObject(forKey: AppIconSettings.modeKey)
-        defaults.removeObject(forKey: AppearanceSettings.appearanceModeKey)
-        defaults.removeObject(forKey: settingsFileBackupsDefaultsKey)
-
-        let directoryURL = try makeTemporaryDirectory()
-        defer { try? FileManager.default.removeItem(at: directoryURL) }
-
-        let managedIconURL = directoryURL.appendingPathComponent("icon.json", isDirectory: false)
-        try writeSettingsFile(
-            """
-            {
-              "app": {
-                "appIcon": "automatic"
-              }
-            }
-            """,
-            to: managedIconURL
-        )
-
-        var startObservationCallCount = 0
-        var stopObservationCallCount = 0
-        var imageRequestCount = 0
-        var runtimeIconSetCount = 0
-        var dockTileNotificationCount = 0
-        AppIconSettings.setLiveEnvironmentProviderForTesting {
-            AppIconSettings.Environment(
-                isApplicationFinishedLaunching: { false },
-                imageForMode: { _ in
-                    imageRequestCount += 1
-                    return nil
-                },
-                setApplicationIconImage: { _ in
-                    runtimeIconSetCount += 1
-                },
-                startAppearanceObservation: {
-                    startObservationCallCount += 1
-                },
-                stopAppearanceObservation: {
-                    stopObservationCallCount += 1
-                },
-                notifyDockTilePlugin: {
-                    dockTileNotificationCount += 1
-                }
-            )
-        }
-
-        _ = KeyboardShortcutSettingsFileStore(
-            primaryPath: managedIconURL.path,
-            fallbackPath: nil,
-            startWatching: false
-        )
-
-        XCTAssertEqual(defaults.string(forKey: AppIconSettings.modeKey), AppIconMode.automatic.rawValue)
-
-        let managedAppearanceURL = directoryURL.appendingPathComponent("appearance.json", isDirectory: false)
-        try writeSettingsFile(
-            """
-            {
-              "app": {
-                "appearance": "system"
-              }
-            }
-            """,
-            to: managedAppearanceURL
-        )
-
-        _ = KeyboardShortcutSettingsFileStore(
-            primaryPath: managedAppearanceURL.path,
-            fallbackPath: nil,
-            startWatching: false
-        )
-
-        XCTAssertNil(defaults.object(forKey: AppIconSettings.modeKey))
-        XCTAssertEqual(defaults.string(forKey: AppearanceSettings.appearanceModeKey), AppearanceMode.system.rawValue)
         XCTAssertEqual(startObservationCallCount, 0)
         XCTAssertEqual(stopObservationCallCount, 0)
         XCTAssertEqual(imageRequestCount, 0)
